requirements.txt
aiogram>=3.3.0
motor>=3.3.2
pymongo>=4.6.1
pytz>=2024.1
APScheduler>=3.10.4
python-dateutil>=2.8.2
python-dotenv>=1.0.1 